PythonApplication1/requirements.txt
flask[async]>=2.0.0
telethon>=1.24.0
requests>=2.25.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
python-telegram-bot[webhooks]>=20.0
nest_asyncio>=1.5.0
gunicorn>=20.1.0
